+++ src/cmdClientDoNone.go
@@ -21,10 +21,9 @@ type clientCreateNoneCmd struct {
 	Gcp           clusterCreateCmdGcp    `no-flag:"true"`
 	Docker        clusterCreateCmdDocker `no-flag:"true"`
 	osSelectorCmd
-	ParallelThreads int     `long:"threads" description:"Run on this many nodes in parallel" default:"50"`
-	PriceOnly       bool    `long:"price" description:"Only display price of ownership; do not actually create the cluster"`
-	Owner           string  `long:"owner" description:"AWS/GCP only: create owner tag with this value"`
-	Help            helpCmd `command:"help" subcommands-optional:"true" description:"Print help"`
+	ParallelThreads int    `long:"threads" description:"Run on this many nodes in parallel" default:"50"`
+	PriceOnly       bool   `long:"price" description:"Only display price of ownership; do not actually create the cluster"`
+	Owner           string `long:"owner" description:"AWS/GCP only: create owner tag with this value"`
 }
 
 func (c *clientCreateNoneCmd) isGrow() bool {
@@ -35,8 +34,15 @@ func (c *clientCreateNoneCmd) isGrow() bool {
 }
 
 func (c *clientCreateNoneCmd) Execute(args []string) error {
-	if earlyProcess(args) {
+	if earlyProcessV2(nil, true) {
 		return nil
+	}
+	if inslice.HasString(args, "help") {
+		if a.opts.Config.Backend.Type == "docker" {
+			printHelp("The aerolab command can be optionally followed by '--' and then extra switches that will be passed directory to Docker. Ex: aerolab cluster create -c 2 -n bob -- -v local:remote --device-read-bps=...\n\n")
+		} else {
+			printHelp("")
+		}
 	}
 	if c.PriceOnly && a.opts.Config.Backend.Type == "docker" {
 		return logFatal("Docker backend does not support pricing")
@@ -133,7 +139,7 @@ func (c *clientCreateNoneCmd) createBase(args []string, nt string) (machines []i
 		privileged:      c.Docker.Privileged,
 		network:         c.Docker.NetworkName,
 		exposePorts:     ep,
-		switches:        c.Docker.ExtraFlags,
+		switches:        args,
 		dockerHostname:  !c.NoSetHostname,
 		ami:             c.Aws.AMI,
 		instanceType:    c.Aws.InstanceType,